refactor: move concrete iroh db into iroh
iroh-bytes should be solely about the iroh data interchange protocol
