+++ iroh-bytes/Cargo.toml
@@ -26,7 +26,6 @@ iroh-net = { path = "../iroh-net" }
 multibase = "0.9.1"
 num_cpus = "1.15.0"
 once_cell = "1.17.0"
-portable-atomic = "1"
 postcard = { version = "1", default-features = false, features = ["alloc", "use-std", "experimental-derive"] }
 quinn = "0.10"
 rand = "0.8"
@@ -42,7 +41,6 @@ tokio-util = { version = "0.7", features = ["io-util", "io", "rt"] }
 tracing = "0.1"
 tracing-futures = "0.2.5"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-walkdir = "2"
 
 [dev-dependencies]
 http-body = "0.4.5"
