+++ iroh/src/commands.rs
@@ -7,7 +7,7 @@ use std::{net::SocketAddr, path::PathBuf};
 use anyhow::{Context, Result};
 use clap::{Parser, Subcommand};
 use iroh::rpc_protocol::*;
-use iroh_bytes::{protocol::RequestToken, provider::Ticket, runtime, Hash};
+use iroh_bytes::{protocol::RequestToken, provider::Ticket, util::runtime, Hash};
 use iroh_net::tls::{Keypair, PeerId};
 use quic_rpc::transport::quinn::QuinnConnection;
 use quic_rpc::RpcClient;
@@ -290,7 +290,7 @@ pub fn create_quinn_client(
 #[cfg(feature = "metrics")]
 pub fn init_metrics_collection(
     metrics_addr: Option<SocketAddr>,
-    rt: &iroh_bytes::runtime::Handle,
+    rt: &iroh_bytes::util::runtime::Handle,
 ) -> Option<tokio::task::JoinHandle<()>> {
     use iroh_metrics::core::Metric;
