+++ iroh/src/util/progress.rs
@@ -16,12 +16,11 @@ use std::sync::Arc;
 use std::task::Poll;
 
 use bytes::Bytes;
+use iroh_bytes::util::io::TrackingWriter;
 use iroh_io::AsyncSliceWriter;
 use portable_atomic::{AtomicU16, AtomicU64};
 use tokio::io::{self, AsyncRead, AsyncWrite};
 use tokio::sync::{broadcast, mpsc};
-
-use super::io::TrackingWriter;
 
 /// A generic progress event emitter.
 ///
