+++ iroh/tests/cli.rs
@@ -9,7 +9,7 @@ use std::str::FromStr;
 use anyhow::{Context, Result};
 use duct::{cmd, ReaderHandle};
 use iroh::bytes::{provider::Ticket, Hash};
-use iroh::database::create_collection;
+use iroh::database::{create_collection, create_data_sources, DbEntry};
 use rand::{RngCore, SeedableRng};
 use regex::Regex;
 use testdir::testdir;
@@ -30,8 +30,6 @@ fn make_rand_file(size: usize, path: &Path) -> Result<Hash> {
 ///
 /// Takes all files and splits them in half, and leaves the collection alone.
 fn make_partial_download(out_dir: &Path) -> anyhow::Result<Hash> {
-    use iroh_bytes::provider::{create_data_sources, DbEntry};
-
     let temp_dir = out_dir.join(".iroh-tmp");
     anyhow::ensure!(!temp_dir.exists());
     std::fs::create_dir_all(&temp_dir)?;
